refactor: Modularize JavaScript application
Split the monolithic script.js into multiple ES6 modules to improve code organization, maintainability, and scalability.

- Create a new `js/modules` directory for the individual modules (Config, State, Utils, API, UI, PWA, Handlers, etc.).
- Create a new main entry point, `js/app.js`, which imports all modules and initializes the application.
- Update `functions.php` to dequeue the old script and enqueue the new `js/app.js` with the `type="module"` attribute, ensuring correct browser loading.
- The original `script.js` has been removed after being used as a reference.
