+++ ting-tong-theme/js/modules/config.js
@@ -1,0 +1,333 @@
+export const Config = {
+  PREFETCH_NEIGHBORS: true,
+  PREFETCH_MARGIN: "150%",
+  UNLOAD_FAR_SLIDES: true,
+  FAR_DISTANCE: 2,
+  LOW_DATA_MODE: false,
+  METRICS_ENDPOINT: "/api/tt-metrics",
+  DEBUG_PANEL: true,
+  GESTURE_GRACE_PERIOD_MS: 2000,
+  TRANSLATIONS: {
+    pl: {
+      loggedOutText: "Nie masz psychy się zalogować",
+      loggedInText: "Ting Tong",
+      loginSuccess: "Zalogowano pomyślnie!",
+      loginFailed: "Logowanie nie powiodło się. Spróbuj ponownie.",
+      accountHeaderText: "Konto",
+      menuAriaLabel: "Menu",
+      subscribeAriaLabel: "subskrajbować",
+      shareTitle: "Udostępnij",
+      shareAriaLabel: "Udostępnij",
+      shareText: "Szeruj",
+      tipTitle: "Napiwek",
+      tipAriaLabel: "Napiwek",
+      tipText: "Napiwek",
+      languageAriaLabel: "Zmień język",
+      languageText: "PL",
+      subscribeAlert: "Zaloguj się, aby subskrajbować.",
+      likeAlert: "Zaloguj się, aby lajkować.",
+      notificationAlert: "Zaloguj się i bądź na bieżąco.",
+      menuAccessAlert: "Zaloguj się, aby uzyskać dostęp do menu.",
+      logoutSuccess: "Zostałeś wylogowany.",
+      likeError: "Błąd komunikacji z serwerem.",
+      secretTitle: "Ściśle Tajne",
+      secretSubtitleAction: "Zaloguj się,",
+      secretSubtitleRest: " aby odblokować",
+      pwaTitle: "Ściśle Tajne",
+      pwaSubtitleAction: "Pobierz aplikację,",
+      pwaSubtitleRest: " aby zobaczyć",
+      closeAccountAriaLabel: "Zamknij panel konta",
+      accountMenuButton: "Konto",
+      logoutLink: "Wyloguj",
+      profileTab: "Profil",
+      passwordTab: "Hasło",
+      deleteTab: "Usuń konto",
+      loggedInState: "Zalogowany",
+      loggedOutState: "Gość",
+      linkCopied: "Link skopiowany do schowka!",
+      likeAriaLabel: "Polub",
+      notificationAriaLabel: "Powiadomienia",
+      commentsAriaLabel: "Komentarze",
+      commentsModalTitle: "Komentarze",
+      closeCommentsAriaLabel: "Zamknij komentarze",
+      likeAriaLabelWithCount: "Polub. Aktualna liczba polubień: {count}",
+      unlikeAriaLabelWithCount:
+        "Cofnij polubienie. Aktualna liczba polubień: {count}",
+      notificationsTitle: "Powiadomienia",
+      closeNotificationsAriaLabel: "Zamknij powiadomienia",
+      notificationsEmpty: "Wszystko na bieżąco!",
+      notif1Preview: "Nowa wiadomość od Admina",
+      notif1Time: "2 min temu",
+      notif1Full:
+        "Cześć! Chcieliśmy tylko dać znać, że nowa wersja aplikacji jest już dostępna. Sprawdź nowe funkcje w panelu konta!",
+      notif2Preview: "Twój profil został zaktualizowany",
+      notif2Time: "10 min temu",
+      notif2Full:
+        "Twoje zmiany w profilu zostały pomyślnie zapisane. Możesz je przejrzeć w dowolnym momencie, klikając w swój awatar.",
+      notif3Preview: "Specjalna oferta czeka na Ciebie!",
+      notif3Time: "1 godz. temu",
+      notif3Full:
+        "Nie przegap! Przygotowaliśmy dla Ciebie specjalną letnią promocję. Zgarnij dodatkowe bonusy już teraz. Oferta ograniczona czasowo.",
+      pwaModalTitle: "Pełne doświadczenie Ting Tong na Twoim telefonie!",
+      pwaModalBody:
+        "Zeskanuj kod QR lub odwiedź nas na telefonie, aby pobrać aplikację i odblokować pełne możliwości.",
+      installPwaHeading: "Zobacz więcej!",
+      installPwaSubheadingAction: "Pobierz aplikację",
+      installPwaSubheadingRest: " i zobacz więcej!",
+      installPwaAction: "Zainstaluj",
+      openPwaAction: "Otwórz",
+      videoErrorTitle: "Błąd Wideo",
+      videoErrorSubtitle: "Nie można załadować materiału.",
+      videoErrorRetry: "Spróbuj ponownie",
+      alreadyInstalledText: "Przecież już ściągłeś",
+      noComments: "Brak komentarzy. Bądź pierwszy!",
+      commentSortTriggerText: "Sortuj według: ",
+      commentSortNewest: "Fresz",
+      commentSortBest: "Best",
+      toggleRepliesShow: "Rozwiń odpowiedzi ({count})",
+      toggleRepliesHide: "Ukryj odpowiedzi",
+      loginToCommentAction: "Zaloguj się",
+      loginToCommentRest: ", aby dodać komentarz.",
+      addCommentPlaceholder: "Dodaj komentarz...",
+      sendCommentAriaLabel: "Wyślij komentarz",
+      commentReplyAction: "Reply",
+      replyingTo: "Odpowiadasz @{user}",
+      cancelReplyAriaLabel: "Anuluj odpowiedź",
+      failedToUpdateLike: "Nie udało się zaktualizować polubienia.",
+      accountModalTitleProfile: "Profil",
+      accountModalTitlePassword: "Hasło",
+      accountModalTitleDelete: "Usuń konto",
+      avatarEditBtnTitle: "Zmień avatar",
+      patronBadgeText: "Patron Miłości",
+      personalDataHeader: "Dane osobowe",
+      firstNameLabel: "Imię",
+      firstNamePlaceholder: "Twoje imię",
+      lastNameLabel: "Nazwisko",
+      lastNamePlaceholder: "Twoje nazwisko",
+      emailLabel: "Email",
+      emailPlaceholder: "email@domena.pl",
+      saveProfileBtn: "Zapisz zmiany",
+      settingsHeader: "Ustawienia",
+      emailConsentLabel: "Zgoda na maile",
+      emailLanguageLabel: "Język maili",
+      emailLangPolish: "Polski",
+      emailLangEnglish: "English",
+      saveSettingsBtn: "Zapisz ustawienia",
+      changePasswordHeader: "Zmiana hasła",
+      currentPasswordLabel: "Obecne hasło",
+      currentPasswordPlaceholder: "Wprowadź obecne hasło",
+      newPasswordLabel: "Nowe hasło",
+      newPasswordPlaceholder: "Minimum 8 znaków",
+      confirmPasswordLabel: "Powtórz nowe hasło",
+      confirmPasswordPlaceholder: "Powtórz nowe hasło",
+      passwordHelperText:
+        "Hasło musi zawierać minimum 8 znaków. Zalecamy użycie liter, cyfr i znaków specjalnych.",
+      changePasswordBtn: "Zmień hasło",
+      deleteAccountHeader: "Usuń konto",
+      deleteWarningHeader: "⚠️ Uwaga!",
+      deleteWarningBody:
+        "Ta operacja jest nieodwracalna. Wszystkie Twoje dane, filmy i ustawienia zostaną trwale usunięte.",
+      deleteConfirmationLabel:
+        "Aby potwierdzić, wpisz: <strong>USUWAM KONTO</strong>",
+      deleteConfirmationPlaceholder: "USUWAM KONTO",
+      deleteHelperText:
+        "Po usunięciu konta zostaniesz automatycznie wylogowany.",
+      deleteAccountBtn: "Trwale usuń konto",
+      profileFollowingLabel: "Obserwuje",
+      profileFollowersLabel: "Obserwujący",
+      profileLikesLabel: "Polubienia",
+      profileFollowBtn: "Obserwuj",
+      savingButtonText: "Zapisywanie...",
+      changingButtonText: "Zmienianie...",
+      deletingButtonText: "Usuwanie...",
+      settingsUpdateSuccess: "Ustawienia zostały zapisane! (DEMO)",
+      settingsUpdateError: "Wystąpił błąd podczas zapisywania ustawień.",
+      fileSelectImageError: "Proszę wybrać plik obrazu.",
+      fileTooLargeError: "Plik jest za duży. Maksymalny rozmiar to 5MB.",
+      avatarUpdateSuccess: "Avatar został zaktualizowany!",
+      avatarUpdateError: "Nie otrzymano URL avatara",
+      imageProcessingError: "Błąd podczas przetwarzania obrazu.",
+      allFieldsRequiredError: "Wszystkie pola są wymagane.",
+      invalidEmailError: "Podaj prawidłowy adres email.",
+      profileUpdateSuccess: "Profil został zaktualizowany!",
+      profileUpdateError: "Wystąpił błąd podczas aktualizacji profilu.",
+      profileUpdateFailedError: "Błąd aktualizacji profilu.",
+      passwordUpdateSuccess: "Hasło zostało zmienione!",
+      passwordUpdateError: "Błąd zmiany hasła.",
+      passwordLengthError: "Nowe hasło musi mieć minimum 8 znaków.",
+      passwordsMismatchError: "Nowe hasła muszą być identyczne.",
+      passwordChangeFailedError: "Błąd zmiany hasła.",
+      deleteAccountSuccess: "Konto zostało usunięte. Trwa wylogowywanie...",
+      deleteAccountError: "Błąd usuwania konta.",
+      deleteAccountFailedError: "Błąd usuwania konta.",
+      deleteConfirmationError: "Wpisz dokładnie: {confirmationText}",
+      deleteConfirmationString: "USUWAM KONTO",
+      postCommentError: "Nie udało się opublikować komentarza.",
+      commentEditAction: "Edytuj",
+      commentDeleteAction: "Usuń",
+      deleteCommentConfirm: "Czy na pewno chcesz usunąć ten komentarz?",
+      editCommentPrompt: "Edytuj swój komentarz:",
+      commentUpdateSuccess: "Komentarz zaktualizowany.",
+      commentUpdateError: "Nie udało się zaktualizować komentarza.",
+      commentDeleteSuccess: "Komentarz usunięty.",
+      commentDeleteError: "Nie udało się usunąć komentarza.",
+      postCommentSuccess: "Komentarz został opublikowany!",
+    },
+    en: {
+      loggedOutText: "You don't have the guts to log in",
+      loggedInText: "You are logged in",
+      loginSuccess: "Logged in successfully!",
+      loginFailed: "Login failed. Please try again.",
+      accountHeaderText: "Account",
+      menuAriaLabel: "Menu",
+      subscribeAriaLabel: "Subscribe",
+      shareTitle: "Share",
+      shareAriaLabel: "Share",
+      shareText: "Share",
+      tipTitle: "Tip",
+      tipAriaLabel: "Tip",
+      tipText: "Tip",
+      languageAriaLabel: "Change language",
+      languageText: "EN",
+      subscribeAlert: "Log in to subscribe.",
+      profileViewAlert: "Log in to see the profile.",
+      likeAlert: "Log in to like.",
+      notificationAlert: "Log in to stay up to date.",
+      menuAccessAlert: "Log in to access the menu.",
+      logoutSuccess: "You have been logged out.",
+      likeError: "Server communication error.",
+      secretTitle: "Top Secret",
+      secretSubtitleAction: "Log in",
+      secretSubtitleRest: " to unlock",
+      pwaTitle: "Top Secret",
+      pwaSubtitleAction: "Download the app",
+      pwaSubtitleRest: " to view",
+      closeAccountAriaLabel: "Close account panel",
+      accountMenuButton: "Account",
+      logoutLink: "Logout",
+      profileTab: "Profile",
+      passwordTab: "Password",
+      deleteTab: "Delete account",
+      loggedInState: "Logged In",
+      loggedOutState: "Guest",
+      linkCopied: "Link copied to clipboard!",
+      likeAriaLabel: "Like",
+      notificationAriaLabel: "Notifications",
+      commentsAriaLabel: "Comments",
+      commentsModalTitle: "Comments",
+      closeCommentsAriaLabel: "Close comments",
+      likeAriaLabelWithCount: "Like. Current likes: {count}",
+      unlikeAriaLabelWithCount: "Unlike. Current likes: {count}",
+      notificationsTitle: "Notifications",
+      closeNotificationsAriaLabel: "Close notifications",
+      notificationsEmpty: "You are all caught up!",
+      notif1Preview: "New message from Admin",
+      notif1Time: "2 mins ago",
+      notif1Full:
+        "Hi there! Just wanted to let you know that a new version of the app is available. Check out the new features in your account panel!",
+      notif2Preview: "Your profile has been updated",
+      notif2Time: "10 mins ago",
+      notif2Full:
+        "Your profile changes have been saved successfully. You can review them anytime by clicking on your avatar.",
+      notif3Preview: "A special offer is waiting for you!",
+      notif3Time: "1 hour ago",
+      notif3Full:
+        "Don't miss out! We have prepared a special summer promotion just for you. Grab your extra bonuses now. Limited time offer.",
+      pwaModalTitle: "The full Ting Tong experience is on your phone!",
+      pwaModalBody:
+        "Scan the QR code below or visit us on your phone to download the app and unlock the full experience.",
+      installPwaHeading: "See more!",
+      installPwaSubheadingAction: "Download the app",
+      installPwaSubheadingRest: " to see more!",
+      installPwaAction: "Install",
+      openPwaAction: "Open",
+      videoErrorTitle: "Video Error",
+      videoErrorSubtitle: "Could not load the content.",
+      videoErrorRetry: "Try Again",
+      alreadyInstalledText: "You've already installed the app!",
+      noComments: "No comments yet. Be the first!",
+      commentSortTriggerText: "Sort by: ",
+      commentSortNewest: "Newest",
+      commentSortBest: "Best",
+      toggleRepliesShow: "Show replies ({count})",
+      toggleRepliesHide: "Hide replies",
+      loginToCommentAction: "Log in",
+      loginToCommentRest: " to add a comment.",
+      addCommentPlaceholder: "Add a comment...",
+      sendCommentAriaLabel: "Send comment",
+      commentReplyAction: "Reply",
+      replyingTo: "Replying to @{user}",
+      cancelReplyAriaLabel: "Cancel reply",
+      failedToUpdateLike: "Failed to update like.",
+      accountModalTitleProfile: "Profile",
+      accountModalTitlePassword: "Password",
+      accountModalTitleDelete: "Delete Account",
+      avatarEditBtnTitle: "Change avatar",
+      patronBadgeText: "Patron of Love",
+      personalDataHeader: "Personal Data",
+      firstNameLabel: "First Name",
+      firstNamePlaceholder: "Your first name",
+      lastNameLabel: "Last Name",
+      lastNamePlaceholder: "Your last name",
+      emailLabel: "Email",
+      emailPlaceholder: "email@domain.com",
+      saveProfileBtn: "Save changes",
+      settingsHeader: "Settings",
+      emailConsentLabel: "Email Consent",
+      emailLanguageLabel: "Email Language",
+      emailLangPolish: "Polish",
+      emailLangEnglish: "English",
+      saveSettingsBtn: "Save settings",
+      changePasswordHeader: "Change Password",
+      currentPasswordLabel: "Current Password",
+      currentPasswordPlaceholder: "Enter current password",
+      newPasswordLabel: "New Password",
+      newPasswordPlaceholder: "Minimum 8 characters",
+      confirmPasswordLabel: "Repeat new password",
+      confirmPasswordPlaceholder: "Repeat new password",
+      passwordHelperText:
+        "Password must be at least 8 characters long. We recommend using letters, numbers, and special characters.",
+      changePasswordBtn: "Change password",
+      deleteAccountHeader: "Delete Account",
+      deleteWarningHeader: "⚠️ Warning!",
+      deleteWarningBody:
+        "This operation is irreversible. All your data, videos, and settings will be permanently deleted.",
+      deleteConfirmationLabel:
+        "To confirm, type: <strong>DELETE ACCOUNT</strong>",
+      deleteConfirmationPlaceholder: "DELETE ACCOUNT",
+      deleteHelperText:
+        "After deleting your account, you will be automatically logged out.",
+      deleteAccountBtn: "Permanently delete account",
+      profileFollowingLabel: "Following",
+      profileFollowersLabel: "Followers",
+      profileLikesLabel: "Likes",
+      profileFollowBtn: "Follow",
+      savingButtonText: "Saving...",
+      changingButtonText: "Changing...",
+      deletingButtonText: "Deleting...",
+      settingsUpdateSuccess: "Settings have been saved! (DEMO)",
+      settingsUpdateError: "An error occurred while saving settings.",
+      fileSelectImageError: "Please select an image file.",
+      fileTooLargeError: "File is too large. Maximum size is 5MB.",
+      avatarUpdateSuccess: "Avatar has been updated!",
+      avatarUpdateError: "Avatar URL not received",
+      imageProcessingError: "Error processing image.",
+      allFieldsRequiredError: "All fields are required.",
+      invalidEmailError: "Please provide a valid email address.",
+      profileUpdateSuccess: "Profile has been updated!",
+      profileUpdateError: "An error occurred while updating the profile.",
+      profileUpdateFailedError: "Profile update failed.",
+      passwordUpdateSuccess: "Password has been changed!",
+      passwordUpdateError: "Error changing password.",
+      passwordLengthError: "New password must be at least 8 characters long.",
+      passwordsMismatchError: "New passwords must be identical.",
+      passwordChangeFailedError: "Password change failed.",
+      deleteAccountSuccess: "Account has been deleted. Logging you out...",
+      deleteAccountError: "Error deleting account.",
+      deleteAccountFailedError: "Account deletion failed.",
+      deleteConfirmationError: "Please type exactly: {confirmationText}",
+      deleteConfirmationString: "DELETE ACCOUNT",
+      postCommentSuccess: "Comment published successfully!",
+    },
+  },
+};
